Подкутил кнопки надо разобраться

+++ internal/config/config.go
@@ -9,7 +9,7 @@ import (
 
 type Config struct {
 	TelegramToken   string
-	TelegramChatId  int
+	TelegramChatId  int64
 	QuoteAPIURL     string
 	TranslateAPIURL string
 }
@@ -27,7 +27,7 @@ func NewConfig() *Config {
 	tgChatId, _ := strconv.Atoi(chatIdStr)
 	return &Config{
 		TelegramToken:   token,
-		TelegramChatId:  tgChatId,
+		TelegramChatId:  int64(tgChatId),
 		QuoteAPIURL:     os.Getenv("QUOTE_API_URL"),
 		TranslateAPIURL: os.Getenv("TRANSLATE_API_URL"),
 	}
